:recycle: s3-integration refactor SupportedFileExtensions
