+++ refarch-integrations/refarch-s3-integration/refarch-s3-integration-client-starter/src/main/java/de/muenchen/refarch/s3/integration/client/configuration/S3IntegrationClientAutoConfiguration.java
@@ -3,8 +3,8 @@ package de.muenchen.refarch.s3.integration.client.configuration;
 import de.muenchen.refarch.s3.integration.client.ApiClient;
 import de.muenchen.refarch.s3.integration.client.api.FileApiApi;
 import de.muenchen.refarch.s3.integration.client.api.FolderApiApi;
+import de.muenchen.refarch.s3.integration.client.domain.model.SupportedFileExtensions;
 import de.muenchen.refarch.s3.integration.client.properties.S3IntegrationClientProperties;
-import de.muenchen.refarch.s3.integration.client.properties.SupportedFileExtensions;
 import de.muenchen.refarch.s3.integration.client.service.ApiClientFactory;
 import de.muenchen.refarch.s3.integration.client.service.FileService;
 import de.muenchen.refarch.s3.integration.client.service.S3DomainProvider;
